requirements.txt
torch
torchvision
Pillow
pytest
